Audio mixer and hitsounds, result screen, minor refactoring

+++ crates/apex-client/src/client/state/gameplay_state.rs
@@ -3,13 +3,13 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct GameplayState {
   /// Offset of the audio in milliseconds
-  pub audio_offset: i32,
+  pub audio_offset: i64,
 
   /// Additional time before the first note
-  pub lead_in: i32,
+  pub lead_in: u64,
 
   /// Additional time after the last note
-  pub lead_out: i32,
+  pub lead_out: u64,
 }
 
 impl Default for GameplayState {
